refactoring location of files
