+++ src/main/java/com/securian/creditcompass/useCases/claimPage/ClaimPageController.java
@@ -1,17 +1,10 @@
-package com.securian.creditcompass.claimPage;
+package com.securian.creditcompass.useCases.claimPage;
 
-import com.securian.creditcompass.orderCalculator.OrderCalculator;
 import com.securian.creditcompass.entities.Claim;
-import com.securian.creditcompass.dataAccess.ClaimRepository;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.server.ResponseStatusException;
-import java.util.List;
-import java.util.ArrayList;
-import java.util.Optional;
 
 @RestController
 public class ClaimPageController {
