+++ src/main/java/com/securian/creditcompass/useCases/claimPage/ClaimPageInputData.java
@@ -1,4 +1,4 @@
-package com.securian.creditcompass.claimPage;
+package com.securian.creditcompass.useCases.claimPage;
 
 import lombok.Getter;
