+++ src/main/java/com/securian/creditcompass/useCases/login/LoginInputData.java
@@ -1,4 +1,4 @@
-package com.securian.creditcompass.login;
+package com.securian.creditcompass.useCases.login;
 
 import lombok.Getter;
